refactor(api): split query_parser.go into 5 files by parsing concern
Split query_parser.go (1748 lines) into focused files:
- query_parser.go: types, Parse(), ParseWithOptions()
- query_parser_select.go: select field and aggregation parsing
- query_parser_order.go: order by and vector order parsing
- query_parser_filter.go: filter, logical, nested filter parsing
- query_parser_sql.go: SQL generation, buildWhereClause, filterToSQL
